Makes legacy event class private and renames internal event class.
