+++ financial-connections/src/test/java/com/stripe/android/financialconnections/TestFinancialConnectionsAnalyticsTracker.kt
@@ -1,14 +1,14 @@
 package com.stripe.android.financialconnections
 
 import com.google.common.truth.Truth
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsAnalyticsEvent
 import com.stripe.android.financialconnections.analytics.FinancialConnectionsAnalyticsTracker
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent
 
 internal class TestFinancialConnectionsAnalyticsTracker : FinancialConnectionsAnalyticsTracker {
 
-    val sentEvents = mutableListOf<FinancialConnectionsInternalEvent>()
+    val sentEvents = mutableListOf<FinancialConnectionsAnalyticsEvent>()
 
-    override suspend fun track(event: FinancialConnectionsInternalEvent): Result<Unit> {
+    override suspend fun track(event: FinancialConnectionsAnalyticsEvent): Result<Unit> {
         sentEvents += event
         return Result.success(Unit)
     }
